third commit
Added instance variable to Block regarding size of the Object to be
drawn. This will also be used to determine the mass of the object
(neasured in units^2).

Added new background and floor, ability to check whether block collides
with wall, and drawing the block on the canvas. Currently there is a bug
in which the drawn block leaves a trail.